added utils for functions
